+++ src/main/java/com/sayjong/backend/auth/dto/TokenRefreshRequestDto.java
@@ -1,4 +1,4 @@
-package com.sayjong.backend.user.dto;
+package com.sayjong.backend.auth.dto;
 
 import lombok.Getter;
 import lombok.Setter;
