+++ src/main/java/com/sayjong/backend/auth/dto/UserSignUpRequestDto.java
@@ -1,11 +1,12 @@
-package com.sayjong.backend.user.dto;
+package com.sayjong.backend.auth.dto;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class UserLoginRequestDto {
+public class UserSignUpRequestDto {
     private String loginId;
     private String userPassword;
+    private String nickname;
 }
