+++ src/main/java/com/sayjong/backend/config/JwtTokenProvider.java
@@ -1,6 +1,6 @@
 package com.sayjong.backend.config;
 
-import com.sayjong.backend.user.dto.TokenInfo;
+import com.sayjong.backend.auth.dto.TokenInfo;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.Keys;
 import jakarta.servlet.http.HttpServletRequest;
@@ -13,7 +13,6 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
 
 import java.nio.charset.StandardCharsets;
 import java.security.Key;
